==> analysis/00_WHO_mortality.R <==
rm(list = ls())

require(tidyverse)

country_codes <- read_csv("data/country_codes")

countries_include <- c(3080, 4010, 4020, 4030, 4038, 4045, 4050, 4055, 4070, 4080, 4085, 
                       4140, 4150, 4170, 4180, 4186, 4188, 4190, 4200, 4210, 
                      4230, 4240, 4270, 4274, 4276, 4280, 4290, 4308)

pop <- read_csv("data/pop", col_types = cols(Admin1 = col_skip(), 
                                              SubDiv = col_skip(), Pop1 = col_skip(), Pop2 = col_double(),
                                             Pop3 = col_double(), Pop4 = col_double(), 
                                             Pop5 = col_double(), Pop6 = col_double(),
                                             Pop7 = col_double(), Pop8 = col_double(),
                                             Pop9 = col_double(), Pop10 = col_double(),
                                             Pop10 = col_double(), Pop11 = col_double(),
                                             Pop12 = col_double(), Pop13 = col_double(),
                                             Pop14 = col_double(), Pop15 = col_double(),
                                             Pop16 = col_double(), Pop17 = col_double(),
                                             Pop18 = col_double(), Pop19 = col_double(),
                                             Pop20 = col_double(), Pop21 = col_double(),
                                             Pop22 = col_double(),Pop23 = col_double(),
                                             Pop24 = col_double(),Pop25 = col_double(), 
                                             Pop26 = col_skip(), Lb = col_skip())) %>%
  filter(Country %in% countries_include & Frmat %in% c('00', '01', '02')) %>%
  pivot_longer(Pop2:Pop25, names_to = "age", values_to = "pop") %>%
  mutate(age = case_when(age == 'Pop2' ~ '0',
                         Frmat %in% c('00','01') & age == 'Pop3' ~ '1-4', 
                         Frmat == '02' & age == 'Pop3' ~ '1-4', 
                         age == 'Pop4' ~ '1-4',
                         age == 'Pop5' ~ '1-4',
                         age == 'Pop6' ~ '1-4',
                         age == 'Pop7' ~ '5-9',
                         age == 'Pop8' ~ '10-14',
                         age == 'Pop9' ~ '15-19',
                         age == 'Pop10' ~ '20-24',
                         age == 'Pop11' ~ '25-29',
                         age == 'Pop12' ~ '30-34',
                         age == 'Pop13' ~ '35-39',
                         age == 'Pop14' ~ '40-44',
                         age == 'Pop15' ~ '45-49',
                         age == 'Pop16' ~ '50-54',
                         age == 'Pop17' ~ '55-59',
                         age == 'Pop18' ~ '60-64',
                         age == 'Pop19' ~ '65-69',
                         age == 'Pop20' ~ '70-74',
                         age == 'Pop21' ~ '75-79',
                         age == 'Pop22' ~ '80-84',
                         Frmat == '00' & age == 'Pop23' ~ '85+',
                         Frmat %in% c('01', '02') & age == 'Pop23' ~ '85+',
                         age == 'Pop24' ~ '85+',
                         age == 'Pop25' ~ '85+')) %>% 
  group_by(Country, Year, Sex, age) %>%
  summarise(pop = sum(pop, na.rm = T))
  

mortality <- rbind(read_csv("data/Morticd10_part1", col_types = cols(Admin1 = col_skip(), 
                                                                     SubDiv = col_skip(), List = col_character(), 
                                                                     Cause = col_character(), IM_Frmat = col_skip(), 
                                                                     Deaths1 = col_skip(), Deaths2 = col_integer(),
                                                                     Deaths3 = col_integer(), Deaths4 = col_integer(), 
                                                                     Deaths5 = col_integer(), Deaths6 = col_integer(),
                                                                     Deaths7 = col_integer(), Deaths8 = col_integer(),
                                                                     Deaths9 = col_integer(), Deaths10 = col_integer(),
                                                                     Deaths10 = col_integer(), Deaths11 = col_integer(),
                                                                     Deaths12 = col_integer(), Deaths13 = col_integer(),
                                                                     Deaths14 = col_integer(), Deaths15 = col_integer(),
                                                                     Deaths16 = col_integer(), Deaths17 = col_integer(),
                                                                     Deaths18 = col_integer(), Deaths19 = col_integer(),
                                                                     Deaths20 = col_integer(), Deaths21 = col_integer(),
                                                                     Deaths22 = col_integer(),Deaths23 = col_integer(),
                                                                     Deaths24 = col_integer(),Deaths25 = col_integer(), 
                                                                     Deaths26 = col_skip(), IM_Deaths1 = col_skip(), 
                                                                     IM_Deaths2 = col_skip(), IM_Deaths3 = col_skip(), 
                                                                     IM_Deaths4 = col_skip())), 
                   read_csv("data/Morticd10_part2", col_types = cols(Admin1 = col_skip(), 
                                                                     SubDiv = col_skip(), List = col_character(), 
                                                                     Cause = col_character(), IM_Frmat = col_skip(), 
                                                                     Deaths1 = col_skip(), Deaths2 = col_integer(),
                                                                     Deaths3 = col_integer(), Deaths4 = col_integer(), 
                                                                     Deaths5 = col_integer(), Deaths6 = col_integer(),
                                                                     Deaths7 = col_integer(), Deaths8 = col_integer(),
                                                                     Deaths9 = col_integer(), Deaths10 = col_integer(),
                                                                     Deaths10 = col_integer(), Deaths11 = col_integer(),
                                                                     Deaths12 = col_integer(), Deaths13 = col_integer(),
                                                                     Deaths14 = col_integer(), Deaths15 = col_integer(),
                                                                     Deaths16 = col_integer(), Deaths17 = col_integer(),
                                                                     Deaths18 = col_integer(), Deaths19 = col_integer(),
                                                                     Deaths20 = col_integer(), Deaths21 = col_integer(),
                                                                     Deaths22 = col_integer(),Deaths23 = col_integer(),
                                                                     Deaths24 = col_integer(),Deaths25 = col_integer(), 
                                                                     Deaths26 = col_skip(), IM_Deaths1 = col_skip(), 
                                                                     IM_Deaths2 = col_skip(), IM_Deaths3 = col_skip(), 
                                                                     IM_Deaths4 = col_skip())),
                   read_csv("data/Morticd10_part3", col_types = cols(Admin1 = col_skip(), 
                                                                      SubDiv = col_skip(), List = col_character(), 
                                                                      Cause = col_character(), IM_Frmat = col_skip(), 
                                                                      Deaths1 = col_skip(), Deaths2 = col_integer(),
                                                                      Deaths3 = col_integer(), Deaths4 = col_integer(), 
                                                                      Deaths5 = col_integer(), Deaths6 = col_integer(),
                                                                      Deaths7 = col_integer(), Deaths8 = col_integer(),
                                                                      Deaths9 = col_integer(), Deaths10 = col_integer(),
                                                                      Deaths10 = col_integer(), Deaths11 = col_integer(),
                                                                      Deaths12 = col_integer(), Deaths13 = col_integer(),
                                                                      Deaths14 = col_integer(), Deaths15 = col_integer(),
                                                                      Deaths16 = col_integer(), Deaths17 = col_integer(),
                                                                      Deaths18 = col_integer(), Deaths19 = col_integer(),
                                                                      Deaths20 = col_integer(), Deaths21 = col_integer(),
                                                                      Deaths22 = col_integer(),Deaths23 = col_integer(),
                                                                      Deaths24 = col_integer(),Deaths25 = col_integer(), 
                                                                      Deaths26 = col_skip(), IM_Deaths1 = col_skip(), 
                                                                      IM_Deaths2 = col_skip(), IM_Deaths3 = col_skip(), 
                                                                      IM_Deaths4 = col_skip())), 
                   read_csv("data/Morticd10_part4", col_types = cols(Admin1 = col_skip(), 
                                                                      SubDiv = col_skip(), List = col_character(), 
                                                                      Cause = col_character(), IM_Frmat = col_skip(), 
                                                                      Deaths1 = col_skip(), Deaths2 = col_integer(),
                                                                      Deaths3 = col_integer(), Deaths4 = col_integer(), 
                                                                      Deaths5 = col_integer(), Deaths6 = col_integer(),
                                                                      Deaths7 = col_integer(), Deaths8 = col_integer(),
                                                                      Deaths9 = col_integer(), Deaths10 = col_integer(),
                                                                      Deaths10 = col_integer(), Deaths11 = col_integer(),
                                                                      Deaths12 = col_integer(), Deaths13 = col_integer(),
                                                                      Deaths14 = col_integer(), Deaths15 = col_integer(),
                                                                      Deaths16 = col_integer(), Deaths17 = col_integer(),
                                                                      Deaths18 = col_integer(), Deaths19 = col_integer(),
                                                                      Deaths20 = col_integer(), Deaths21 = col_integer(),
                                                                      Deaths22 = col_integer(),Deaths23 = col_integer(),
                                                                      Deaths24 = col_integer(),Deaths25 = col_integer(), 
                                                                      Deaths26 = col_skip(), IM_Deaths1 = col_skip(), 
                                                                      IM_Deaths2 = col_skip(), IM_Deaths3 = col_skip(), 
                                                                      IM_Deaths4 = col_skip())), 
                   read_csv("data/Morticd10_part5", col_types = cols(Admin1 = col_skip(), 
                                                                      SubDiv = col_skip(), List = col_character(), 
                                                                      Cause = col_character(), IM_Frmat = col_skip(), 
                                                                      Deaths1 = col_skip(), Deaths2 = col_integer(),
                                                                      Deaths3 = col_integer(), Deaths4 = col_integer(), 
                                                                      Deaths5 = col_integer(), Deaths6 = col_integer(),
                                                                      Deaths7 = col_integer(), Deaths8 = col_integer(),
                                                                      Deaths9 = col_integer(), Deaths10 = col_integer(),
                                                                      Deaths10 = col_integer(), Deaths11 = col_integer(),
                                                                      Deaths12 = col_integer(), Deaths13 = col_integer(),
                                                                      Deaths14 = col_integer(), Deaths15 = col_integer(),
                                                                      Deaths16 = col_integer(), Deaths17 = col_integer(),
                                                                      Deaths18 = col_integer(), Deaths19 = col_integer(),
                                                                      Deaths20 = col_integer(), Deaths21 = col_integer(),
                                                                      Deaths22 = col_integer(),Deaths23 = col_integer(),
                                                                      Deaths24 = col_integer(),Deaths25 = col_integer(), 
                                                                      Deaths26 = col_skip(), IM_Deaths1 = col_skip(), 
                                                                      IM_Deaths2 = col_skip(), IM_Deaths3 = col_skip(), 
                                                                      IM_Deaths4 = col_skip()))) %>%
  filter(Country %in% countries_include & Frmat %in% c('00', '01', '02') & List %in% c('103', '104') & Sex %in% c(1,2)) %>%
  pivot_longer(Deaths2:Deaths25, names_to = "age", values_to = "deaths") %>%
  mutate(age = case_when(age == 'Deaths2' ~ '0',
                         Frmat %in% c('00','01') & age == 'Deaths3' ~ '1-4', 
                         Frmat == '02' & age == 'Deaths3' ~ '1-4', 
                         age == 'Deaths4' ~ '1-4',
                         age == 'Deaths5' ~ '1-4',
                         age == 'Deaths6' ~ '1-4',
                         age == 'Deaths7' ~ '5-9',
                         age == 'Deaths8' ~ '10-14',
                         age == 'Deaths9' ~ '15-19',
                         age == 'Deaths10' ~ '20-24',
                         age == 'Deaths11' ~ '25-29',
                         age == 'Deaths12' ~ '30-34',
                         age == 'Deaths13' ~ '35-39',
                         age == 'Deaths14' ~ '40-44',
                         age == 'Deaths15' ~ '45-49',
                         age == 'Deaths16' ~ '50-54',
                         age == 'Deaths17' ~ '55-59',
                         age == 'Deaths18' ~ '60-64',
                         age == 'Deaths19' ~ '65-69',
                         age == 'Deaths20' ~ '70-74',
                         age == 'Deaths21' ~ '75-79',
                         age == 'Deaths22' ~ '80-84',
                         Frmat == '00' & age == 'Deaths23' ~ '85+',
                         Frmat %in% c('01', '02') & age == 'Deaths23' ~ '85+',
                         age == 'Deaths24' ~ '85+',
                         age == 'Deaths25' ~ '85+')) %>%
  group_by(Country, Year, Sex, age, List, Cause) %>%
  summarise(deaths = sum(deaths, na.rm = T))


data <- left_join(mortality, pop, by = c('Country', 'Year',  'Sex', 'age')) %>%
  left_join(., country_codes, by = c('Country' = 'country')) %>%
  ungroup() %>%
  select(-Country) %>%
  rename(country = name) %>%
  rename_all(tolower) %>%
  filter(!is.na(deaths) & !is.na(pop))

data = data %>%
  filter(year %in% 2005:2019) %>%

saveRDS(data, 'temp/ICD10.rds')
